requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain-core
langchain-openai
langgraph
openai
numpy
tqdm
mcp
pyyaml
